Update the Utils folder, add function folder which contains fetchData.js
